+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 1.x branch
+## 1.7 branch
+# 1.7.0
+* `ThumberShell` has been replaced with console commands. Every method of the
+    previous class is now a `Command` class.
+
 ## 1.6 branch
 # 1.6.1
 * added `ConsoleIntegrationTestCase` class.